pyproject.toml
[project]
name = "agent-runner"
version = "0.1.0"
description = "Local-first, agent-agnostic orchestration tool for running AI personas."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Agent-Runner contributors" }]
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.2",
  "rich>=13.9.4",
]

[project.scripts]
agent-runner = "agent_runner.cli:app"

[tool.uv]
# uv will manage the environment; lockfile is generated by `uv sync`.

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
